// multivendor-admin/lib/utils/interfaces/forms/food.form.interface.ts
import { IDropdownSelectItem } from '../global.interface';

export interface IFoodErrors {
  title: string[];
  description: string[];
  image: string[];
  category: string[];
  subCategory: string[];
}

export interface IFoodDetailsForm {
  _id: string | null;
  title: string;
  description: string;
  image: string;
  category: IDropdownSelectItem | null;
  subCategory: IDropdownSelectItem | null;
}